requirements.txt
Flask
firebase-admin
gunicorn
watchdog
Flask-Mail
python-dotenv